pyproject.toml
[tool.poetry]
name = "automl-pipeline"
version = "1.0.0"
description = "ML-пайплайн: мониторинг data drift → AutoML (PyCaret) → MLflow Registry → A/B тест (Flask)"
authors = ["Ermakov Nikita <nikita1234563915@gmail.com>"]
readme = "README.md"
packages = [
    { include = "scripts" },
    { include = "dags" },
]

[tool.poetry.dependencies]
python = "^3.10,<3.12"

# AutoML
pycaret = { version = "^3.3", extras = ["full"] }

# Experiment tracking & model registry
mlflow = "^2.14"

# Orchestration
apache-airflow = "^2.9"

# Web framework (A/B test router)
flask = "^3.0"

# Data drift monitoring
evidently = "^0.4"
scipy = "^1.11"

# Data & ML
pandas = "^2.0"
numpy = "^1.24"
scikit-learn = "^1.3"

# Visualization
matplotlib = "^3.7"
seaborn = "^0.12"

# Utilities
python-dotenv = "^1.0"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
ruff = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
